refactor(gui): lift picker-navigation helpers to _shared
